Add HOST_IP env var to Ambassador deployments
HOST_IP is referenced in a couple locations in our documentation for as
a resolvable env var that allows Ambassador to dynamically target a
service running as a `DaemonSet` (like Consul server or DataDog collector).

This env var is available by default with the YAML deploys but not other
methods. This commit adds it to the other AES deployments

Signed-off-by: Noah Krause <krausenoah@gmail.com>